requirements.txt
cliff==1.4.5
python-circuit
PyYAML==3.10
argparse==1.2.1
distribute==0.6.24
requests>=2.0
#git+https://github.com/gilliam/gilliam-py.git
shortuuid
python-dateutil==2.1
stevedore==0.12
